backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
django-mptt>=0.14,<0.15
cryptography>=42.0,<44.0
django-cors-headers>=4.3,<4.4
django-filter>=24.1,<25.0
PyYAML>=6.0,<7.0
gunicorn>=22.0,<23.0
watchdog>=6.0,<7.0
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=6.0,<7.0
